requirements1.txt
agno==1.1.13
duckduckgo_search==7.5.2
google-genai==1.7.0
pillow==11.1.0
streamlit==1.43.2
python-dotenv==1.1.0
bcrypt==4.3.0
APScheduler==3.11.0
pymongo==4.11.3
deepgram-sdk==3.11.0
PyAudio==0.2.14
sounddevice==0.5.1
soundfile==0.13.1
SpeechRecognition==3.14.2